Decouple User and Review entirely. Actually fix the jest timeout issue.
