--- tests/jest.config.js
@@ -1,5 +1,0 @@
-
-module.exports = {
-    verbose: true,
-    setupFilesAfterEnv: ['./jest.setup.js'],
-};
