--- tests/jest.setup.js
@@ -1,1 +1,0 @@
-jest.setTimeout(10000); // in milliseconds
