+++ tests/jestDefaultTimeout.js
@@ -1,0 +1,1 @@
+jest.setTimeout(30000)
