+++ tests/user/user.test.js
@@ -457,95 +457,6 @@ describe('Testing users with verified accounts', () => {
     }) 
 })
 
-describe("Testing rating system", () => {
-    const username = 'test_reviewee'
-    const testReview1 = new Review({
-        reviewerUsername: 'test_reviewer', 
-        revieweeUsername: 'test_reviewee', 
-        rating: 1, 
-    })
-    const testReview2 = new Review({
-        reviewerUsername: 'test_reviewer', 
-        revieweeUsername: 'test_reviewee', 
-        rating: 2, 
-    })
-    const testReview3 = new Review({
-        reviewerUsername: 'test_reviewer', 
-        revieweeUsername: 'test_reviewee', 
-        rating: 3, 
-    })
-
-    beforeEach(() => {
-        new Review(testReview1).save()
-        new Review(testReview2).save()
-        return new Review(testReview3).save()
-    }) 
-
-    afterEach(() => {
-        return Review.deleteMany() 
-    })
-
-    test("Calculate average rating of a user by aggregating all reviews made to them.", async () => {
-        expect.assertions(1);
-        const expectedRating = ((testReview1.rating + testReview2.rating + testReview3.rating) / 3).toFixed(2)
-        return db.getAverageRating(username).then((rating) => {
-            expect(rating).toBe(expectedRating);
-        })
-    })
-
-    test("When retrieving the average rating of a user without at least 3 ratings, should result in an error message.", async () => {
-        expect.assertions(1)
-        try {
-            await db.getAverageRating('user_without_sufficient_reviews') 
-        }
-        catch(e) {
-            await expect(e).toBe("User must have at least 1 rating(s) to display an average rating!")
-        }
-    })
-
-    test("Get all of the reviews associated with a user with at least 1 review.",  () => {
-        const expectedReviews = [testReview1._id, testReview2._id, testReview3._id]
-        return db.getUserReviews(username).then((reviews) => {
-            expect(reviews.map(a => a._id)).toStrictEqual(expectedReviews)
-        })
-    })
-
-    test("Get all of the reviews associated with a user with 0 reviews.",  () => {
-        return db.getUserReviews('user_that_does_not_exist').then((reviews) => {
-            expect(reviews.length).toBe(0)
-        })
-    })
-
-    describe("Testing API endpoints for rating system", () => {
-        test("When requesting the average rating of a user that is in the database, should return 200 response code", async () => {
-            await request(app)
-                .get("/users/rating")
-                .query({username})
-                .expect(200)
-        })
-
-        test("When requesting the average rating of a user that has no reviews, should return 404 response code", async () => {
-            await request(app)
-                .get("/users/rating")
-                .query({username: 'does_not_exist'})
-                .expect(404)
-        })
-
-        test("When requsting all the reviews left for a user, should expect 200 response code.", async () => {
-            await request(app)
-                .get('/users/reviews') 
-                .query({username})
-                .expect(200)
-        })
-        test("When requsting all the reviews left for a user with no reviews, should expect 200 response code still.", async () => {
-            await request(app)
-                .get('/users/reviews') 
-                .query({username: 'user_that_does_not_exist'})
-                .expect(200)
-        })
-
-    }) 
-})
